requirements.txt
flask
python-dotenv
telethon
requests
